Week2/model and texture (#2)
* Adds manual OBJ

Adds the OBJ file manually (which is hopefully a temporary solution) .

* Adds baseline cube and texture

Adds base cube project (to build) off later. Also adds a texture for the OBJ file.

* Textured cube

Modifies the program to texture the 3D cube using the texture and semi-random texture coordinates.

* Adjusts texture coordinates

Adjusts texture coordinates so that they map correctly on the front side.